Make creation of button images much faster (#2)
Do not use `[UIImage imageNamedRetina:tintColor:]` to tint images anymore. It required to create an image for each available trait collection and highlighted / selected combinations and to redraw the ellipse for each icons instead of once for each size/properties.

Instead, add a new `iconImageView` layer and use `tintColor` along with template rendering mode to tint the icon.